+++ mobile/android/android-components/components/feature/prompts/src/test/java/mozilla/components/feature/prompts/PromptFeatureTest.kt
@@ -8,6 +8,7 @@ import android.app.Activity
 import android.app.Activity.RESULT_CANCELED
 import android.app.Activity.RESULT_OK
 import android.content.ClipData
+import android.content.ContentResolver
 import android.content.Context
 import android.content.Intent
 import android.graphics.Bitmap
@@ -131,6 +132,7 @@ class PromptFeatureTest {
         val feature = spy(
             PromptFeature(
                 fragment = mock(),
+                fileUploadsDirCleaner = mock(),
                 store = store,
                 tabsUseCases = mock(),
                 fragmentManager = fragmentManager,
@@ -150,6 +152,7 @@ class PromptFeatureTest {
                 fragment = mock(),
                 store = store,
                 tabsUseCases = mock(),
+                fileUploadsDirCleaner = mock(),
                 customTabId = "custom-tab",
                 fragmentManager = fragmentManager,
             ) { },
@@ -170,6 +173,7 @@ class PromptFeatureTest {
                 store = store,
                 tabsUseCases = mock(),
                 customTabId = tabId,
+                fileUploadsDirCleaner = mock(),
                 fragmentManager = fragmentManager,
             ) { },
         )
@@ -187,6 +191,7 @@ class PromptFeatureTest {
                 fragment = mock(),
                 store = store,
                 tabsUseCases = mock(),
+                fileUploadsDirCleaner = mock(),
                 fragmentManager = fragmentManager,
             ) { }
         feature.start()
@@ -204,6 +209,7 @@ class PromptFeatureTest {
                 fragment = mock(),
                 tabsUseCases = mock(),
                 store = store,
+                fileUploadsDirCleaner = mock(),
                 fragmentManager = fragmentManager,
             ) { }
         feature.start()
@@ -230,6 +236,7 @@ class PromptFeatureTest {
                 activity = mock(),
                 tabsUseCases = mock(),
                 store = store,
+                fileUploadsDirCleaner = mock(),
                 fragmentManager = fragmentManager,
             ) { }
         feature.start()
@@ -251,6 +258,7 @@ class PromptFeatureTest {
                 activity = mock(),
                 store = store,
                 tabsUseCases = mock(),
+                fileUploadsDirCleaner = mock(),
                 fragmentManager = fragmentManager,
             ) { }
         feature.start()
@@ -279,6 +287,7 @@ class PromptFeatureTest {
                 activity = mock(),
                 store = store,
                 tabsUseCases = mock(),
+                fileUploadsDirCleaner = mock(),
                 fragmentManager = fragmentManager,
             ) { }
         feature.start()
@@ -295,6 +304,7 @@ class PromptFeatureTest {
                 mock<Activity>(),
                 store,
                 tabsUseCases = mock(),
+                fileUploadsDirCleaner = mock(),
                 fragmentManager = fragmentManager,
             ) { },
         )
@@ -314,6 +324,7 @@ class PromptFeatureTest {
                 store,
                 tabsUseCases = mock(),
                 fragmentManager = fragmentManager,
+                fileUploadsDirCleaner = mock(),
                 loginDelegate = object : LoginDelegate {
                     override val loginPickerView = loginPickerView
                     override val onManageLogins = {}
@@ -355,6 +366,7 @@ class PromptFeatureTest {
                 store,
                 tabsUseCases = mock(),
                 fragmentManager = fragmentManager,
+                fileUploadsDirCleaner = mock(),
                 exitFullscreenUsecase = mock(),
                 isSaveLoginEnabled = { true },
                 loginValidationDelegate = mock(),
@@ -387,6 +399,7 @@ class PromptFeatureTest {
             PromptFeature(
                 activity = mock(),
                 store = store,
+                fileUploadsDirCleaner = mock(),
                 tabsUseCases = mock(),
                 fragmentManager = fragmentManager,
                 isSaveLoginEnabled = { false },
@@ -415,6 +428,7 @@ class PromptFeatureTest {
             PromptFeature(
                 activity = mock(),
                 store = store,
+                fileUploadsDirCleaner = mock(),
                 tabsUseCases = mock(),
                 fragmentManager = fragmentManager,
                 isSaveLoginEnabled = { true },
@@ -446,6 +460,7 @@ class PromptFeatureTest {
         val feature = PromptFeature(
             activity = mock(),
             store = store,
+            fileUploadsDirCleaner = mock(),
             tabsUseCases = mock(),
             fragmentManager = fragmentManager,
         ) {}
@@ -475,6 +490,7 @@ class PromptFeatureTest {
                 mock<Activity>(),
                 store,
                 tabsUseCases = mock(),
+                fileUploadsDirCleaner = mock(),
                 fragmentManager = fragmentManager,
                 loginDelegate = object : LoginDelegate {
                     override val loginPickerView = loginPickerView
@@ -505,6 +521,7 @@ class PromptFeatureTest {
                 store,
                 tabsUseCases = mock(),
                 fragmentManager = fragmentManager,
+                fileUploadsDirCleaner = mock(),
                 loginDelegate = object : LoginDelegate {
                     override val loginPickerView = loginPickerView
                     override val onManageLogins = {}
@@ -533,6 +550,7 @@ class PromptFeatureTest {
                 mock<Activity>(),
                 store,
                 tabsUseCases = mock(),
+                fileUploadsDirCleaner = mock(),
                 fragmentManager = fragmentManager,
                 loginDelegate = object : LoginDelegate {
                     override val loginPickerView = loginPickerView
@@ -564,6 +582,7 @@ class PromptFeatureTest {
                 store,
                 tabsUseCases = mock(),
                 fragmentManager = fragmentManager,
+                fileUploadsDirCleaner = mock(),
                 creditCardDelegate = object : CreditCardDelegate {
                     override val creditCardPickerView = creditCardPickerView
                     override val onSelectCreditCard = {}
@@ -593,6 +612,7 @@ class PromptFeatureTest {
                 store,
                 tabsUseCases = mock(),
                 fragmentManager = fragmentManager,
+                fileUploadsDirCleaner = mock(),
                 creditCardDelegate = object : CreditCardDelegate {
                     override val creditCardPickerView = creditCardPickerView
                     override val onSelectCreditCard = {}
@@ -621,6 +641,7 @@ class PromptFeatureTest {
                 store,
                 tabsUseCases = mock(),
                 fragmentManager = fragmentManager,
+                fileUploadsDirCleaner = mock(),
                 creditCardDelegate = object : CreditCardDelegate {
                     override val creditCardPickerView = creditCardPickerView
                     override val onSelectCreditCard = {}
@@ -649,6 +670,7 @@ class PromptFeatureTest {
                 mock<Activity>(),
                 store,
                 tabsUseCases = mock(),
+                fileUploadsDirCleaner = mock(),
                 fragmentManager = fragmentManager,
                 creditCardDelegate = object : CreditCardDelegate {
                     override val creditCardPickerView = creditCardPickerView
@@ -683,6 +705,7 @@ class PromptFeatureTest {
                 mock<Activity>(),
                 store,
                 tabsUseCases = mock(),
+                fileUploadsDirCleaner = mock(),
                 fragmentManager = fragmentManager,
                 addressDelegate = addressDelegate,
             ) { },
@@ -716,6 +739,7 @@ class PromptFeatureTest {
             PromptFeature(
                 mock<Activity>(),
                 store,
+                fileUploadsDirCleaner = mock(),
                 tabsUseCases = mock(),
                 fragmentManager = fragmentManager,
                 addressDelegate = addressDelegate,
@@ -740,6 +764,7 @@ class PromptFeatureTest {
             PromptFeature(
                 activity = mock(),
                 store = store,
+                fileUploadsDirCleaner = mock(),
                 tabsUseCases = mock(),
                 fragmentManager = fragmentManager,
             ) { }
@@ -762,6 +787,7 @@ class PromptFeatureTest {
             PromptFeature(
                 activity = mock(),
                 store = store,
+                fileUploadsDirCleaner = mock(),
                 tabsUseCases = mock(),
                 fragmentManager = fragmentManager,
                 exitFullscreenUsecase = mock(),
@@ -786,6 +812,7 @@ class PromptFeatureTest {
             PromptFeature(
                 activity = mock(),
                 store = store,
+                fileUploadsDirCleaner = mock(),
                 tabsUseCases = mock(),
                 fragmentManager = fragmentManager,
                 exitFullscreenUsecase = mock(),
@@ -810,6 +837,7 @@ class PromptFeatureTest {
             PromptFeature(
                 activity = mock(),
                 store = store,
+                fileUploadsDirCleaner = mock(),
                 tabsUseCases = mock(),
                 fragmentManager = fragmentManager,
                 exitFullscreenUsecase = mock(),
@@ -834,6 +862,7 @@ class PromptFeatureTest {
             PromptFeature(
                 activity = mock(),
                 store = store,
+                fileUploadsDirCleaner = mock(),
                 tabsUseCases = mock(),
                 fragmentManager = fragmentManager,
                 exitFullscreenUsecase = mock(),
@@ -870,6 +899,7 @@ class PromptFeatureTest {
             PromptFeature(
                 activity = mock(),
                 store = store,
+                fileUploadsDirCleaner = mock(),
                 tabsUseCases = mock(),
                 fragmentManager = fragmentManager,
                 exitFullscreenUsecase = mock(),
@@ -891,6 +921,7 @@ class PromptFeatureTest {
         val feature =
             PromptFeature(
                 activity = mock(),
+                fileUploadsDirCleaner = mock(),
                 store = store,
                 tabsUseCases = mock(),
                 fragmentManager = fragmentManager,
@@ -928,6 +959,7 @@ class PromptFeatureTest {
         val feature =
             PromptFeature(
                 activity = mock(),
+                fileUploadsDirCleaner = mock(),
                 store = store,
                 tabsUseCases = mock(),
                 fragmentManager = fragmentManager,
@@ -966,6 +998,7 @@ class PromptFeatureTest {
         timeSelectionTypes.forEach { type ->
             val feature = PromptFeature(
                 activity = mock(),
+                fileUploadsDirCleaner = mock(),
                 store = store,
                 tabsUseCases = mock(),
                 fragmentManager = fragmentManager,
@@ -1009,6 +1042,7 @@ class PromptFeatureTest {
         val feature =
             PromptFeature(
                 activity = mock(),
+                fileUploadsDirCleaner = mock(),
                 tabsUseCases = mock(),
                 store = store,
                 fragmentManager = fragmentManager,
@@ -1026,10 +1060,13 @@ class PromptFeatureTest {
 
         val filePickerRequest =
             PromptRequest.File(emptyArray(), false, onSingleFileSelection, { _, _ -> }) { }
+        val activity = mock<Activity>()
+        doReturn(mock<ContentResolver>()).`when`(activity).contentResolver
 
         val feature =
             PromptFeature(
-                activity = mock(),
+                activity = activity,
+                fileUploadsDirCleaner = mock(),
                 store = store,
                 tabsUseCases = mock(),
                 fragmentManager = fragmentManager,
@@ -1056,10 +1093,13 @@ class PromptFeatureTest {
 
         val filePickerRequest =
             PromptRequest.File(emptyArray(), true, { _, _ -> }, onMultipleFileSelection) {}
+        val activity = mock<Activity>()
+        doReturn(mock<ContentResolver>()).`when`(activity).contentResolver
 
         val feature =
             PromptFeature(
-                activity = mock(),
+                activity = activity,
+                fileUploadsDirCleaner = mock(),
                 tabsUseCases = mock(),
                 store = store,
                 fragmentManager = fragmentManager,
@@ -1098,6 +1138,7 @@ class PromptFeatureTest {
             PromptFeature(
                 activity = mock(),
                 store = store,
+                fileUploadsDirCleaner = mock(),
                 tabsUseCases = mock(),
                 fragmentManager = fragmentManager,
             ) { }
@@ -1120,6 +1161,7 @@ class PromptFeatureTest {
                 activity = mock(),
                 store = store,
                 tabsUseCases = mock(),
+                fileUploadsDirCleaner = mock(),
                 fragmentManager = fragmentManager,
                 creditCardDelegate = object : CreditCardDelegate {
                     override val creditCardPickerView = creditCardPickerView
@@ -1143,6 +1185,7 @@ class PromptFeatureTest {
             PromptFeature(
                 activity = mock(),
                 store = store,
+                fileUploadsDirCleaner = mock(),
                 tabsUseCases = mock(),
                 fragmentManager = fragmentManager,
                 creditCardDelegate = object : CreditCardDelegate {
@@ -1169,6 +1212,7 @@ class PromptFeatureTest {
                 store = store,
                 tabsUseCases = mock(),
                 fragmentManager = fragmentManager,
+                fileUploadsDirCleaner = mock(),
                 creditCardDelegate = object : CreditCardDelegate {
                     override val creditCardPickerView = creditCardPickerView
                     override val onSelectCreditCard = {}
@@ -1191,6 +1235,7 @@ class PromptFeatureTest {
                 activity = mock(),
                 store = store,
                 tabsUseCases = mock(),
+                fileUploadsDirCleaner = mock(),
                 fragmentManager = fragmentManager,
                 creditCardDelegate = object : CreditCardDelegate {
                     override val creditCardPickerView = creditCardPickerView
@@ -1290,6 +1335,7 @@ class PromptFeatureTest {
         val feature =
             PromptFeature(
                 activity = mock(),
+                fileUploadsDirCleaner = mock(),
                 store = store,
                 tabsUseCases = mock(),
                 fragmentManager = fragmentManager,
@@ -1348,6 +1394,7 @@ class PromptFeatureTest {
             PromptFeature(
                 fragment = fragment,
                 store = store,
+                fileUploadsDirCleaner = mock(),
                 tabsUseCases = mock(),
                 fragmentManager = fragmentManager,
                 exitFullscreenUsecase = mock(),
@@ -1376,6 +1423,7 @@ class PromptFeatureTest {
         val feature =
             PromptFeature(
                 activity = mock(),
+                fileUploadsDirCleaner = mock(),
                 store = store,
                 tabsUseCases = mock(),
                 fragmentManager = fragmentManager,
@@ -1414,6 +1462,7 @@ class PromptFeatureTest {
             PromptFeature(
                 activity = mock(),
                 store = store,
+                fileUploadsDirCleaner = mock(),
                 tabsUseCases = mock(),
                 fragmentManager = fragmentManager,
                 exitFullscreenUsecase = mock(),
@@ -1460,6 +1509,7 @@ class PromptFeatureTest {
                 fragment = fragment,
                 store = store,
                 tabsUseCases = mock(),
+                fileUploadsDirCleaner = mock(),
                 fragmentManager = fragmentManager,
                 exitFullscreenUsecase = mock(),
             ) { }
@@ -1491,6 +1541,7 @@ class PromptFeatureTest {
         val feature =
             PromptFeature(
                 fragment = fragment,
+                fileUploadsDirCleaner = mock(),
                 store = store,
                 tabsUseCases = mock(),
                 fragmentManager = fragmentManager,
@@ -1533,6 +1584,7 @@ class PromptFeatureTest {
         val feature =
             PromptFeature(
                 fragment = fragment,
+                fileUploadsDirCleaner = mock(),
                 store = store,
                 tabsUseCases = mock(),
                 fragmentManager = fragmentManager,
@@ -1559,6 +1611,7 @@ class PromptFeatureTest {
         val feature =
             PromptFeature(
                 activity = mock(),
+                fileUploadsDirCleaner = mock(),
                 store = store,
                 tabsUseCases = mock(),
                 fragmentManager = fragmentManager,
@@ -1620,6 +1673,7 @@ class PromptFeatureTest {
         val feature =
             PromptFeature(
                 activity = mock(),
+                fileUploadsDirCleaner = mock(),
                 store = store,
                 tabsUseCases = mock(),
                 fragmentManager = fragmentManager,
@@ -1661,6 +1715,7 @@ class PromptFeatureTest {
         val feature =
             PromptFeature(
                 activity = mock(),
+                fileUploadsDirCleaner = mock(),
                 store = store,
                 tabsUseCases = mock(),
                 fragmentManager = fragmentManager,
@@ -1691,6 +1746,7 @@ class PromptFeatureTest {
         val feature =
             PromptFeature(
                 activity = mock(),
+                fileUploadsDirCleaner = mock(),
                 store = store,
                 tabsUseCases = mock(),
                 fragmentManager = fragmentManager,
@@ -1722,6 +1778,7 @@ class PromptFeatureTest {
         val feature = PromptFeature(
             activity = Robolectric.buildActivity(Activity::class.java).setup().get(),
             store = store,
+            fileUploadsDirCleaner = mock(),
             tabsUseCases = mock(),
             fragmentManager = fragmentManager,
             exitFullscreenUsecase = mock(),
@@ -1754,6 +1811,7 @@ class PromptFeatureTest {
             PromptFeature(
                 activity = mock(),
                 store = store,
+                fileUploadsDirCleaner = mock(),
                 tabsUseCases = mock(),
                 fragmentManager = fragmentManager,
                 exitFullscreenUsecase = mock(),
@@ -1792,6 +1850,7 @@ class PromptFeatureTest {
             PromptFeature(
                 activity = mock(),
                 store = store,
+                fileUploadsDirCleaner = mock(),
                 tabsUseCases = mock(),
                 fragmentManager = fragmentManager,
                 exitFullscreenUsecase = mock(),
@@ -1839,6 +1898,7 @@ class PromptFeatureTest {
             PromptFeature(
                 activity,
                 store,
+                fileUploadsDirCleaner = mock(),
                 tabsUseCases = mock(),
                 customTabId = "custom-tab",
                 shareDelegate = delegate,
@@ -1869,6 +1929,7 @@ class PromptFeatureTest {
                 tabsUseCases = mock(),
                 customTabId = "custom-tab",
                 fragmentManager = fragmentManager,
+                fileUploadsDirCleaner = mock(),
                 isCreditCardAutofillEnabled = { true },
             ) { },
         )
@@ -1889,6 +1950,7 @@ class PromptFeatureTest {
             PromptFeature(
                 mock<Activity>(),
                 store,
+                fileUploadsDirCleaner = mock(),
                 tabsUseCases = mock(),
                 customTabId = "custom-tab",
                 fragmentManager = fragmentManager,
@@ -1912,6 +1974,7 @@ class PromptFeatureTest {
             PromptFeature(
                 mock<Activity>(),
                 store,
+                fileUploadsDirCleaner = mock(),
                 tabsUseCases = mock(),
                 customTabId = "custom-tab",
                 fragmentManager = fragmentManager,
@@ -1935,6 +1998,7 @@ class PromptFeatureTest {
         val feature = PromptFeature(
             fragment = mock(),
             store = store,
+            fileUploadsDirCleaner = mock(),
             tabsUseCases = mock(),
             customTabId = "custom-tab",
             fragmentManager = fragmentManager,
@@ -1953,6 +2017,7 @@ class PromptFeatureTest {
         val exitFullScreenUseCase: SessionUseCases.ExitFullScreenUseCase = mock()
         val feature = PromptFeature(
             fragment = mock(),
+            fileUploadsDirCleaner = mock(),
             store = store,
             tabsUseCases = mock(),
             fragmentManager = fragmentManager,
@@ -1983,6 +2048,7 @@ class PromptFeatureTest {
             fragment = mock(),
             store = store,
             tabsUseCases = mock(),
+            fileUploadsDirCleaner = mock(),
             fragmentManager = fragmentManager,
             exitFullscreenUsecase = exitFullScreenUseCase,
         ) { }
@@ -2017,6 +2083,7 @@ class PromptFeatureTest {
                 store = store,
                 tabsUseCases = mock(),
                 fragmentManager = fragmentManager,
+                fileUploadsDirCleaner = mock(),
                 creditCardValidationDelegate = mock(),
                 isCreditCardAutofillEnabled = { false },
             ) {},
@@ -2053,6 +2120,7 @@ class PromptFeatureTest {
                 store = store,
                 tabsUseCases = mock(),
                 fragmentManager = fragmentManager,
+                fileUploadsDirCleaner = mock(),
                 creditCardValidationDelegate = null,
                 isCreditCardAutofillEnabled = { true },
             ) {},
@@ -2092,6 +2160,7 @@ class PromptFeatureTest {
                 tabsUseCases = mock(),
                 fragmentManager = fragmentManager,
                 creditCardValidationDelegate = mock(),
+                fileUploadsDirCleaner = mock(),
                 isCreditCardAutofillEnabled = { true },
             ) {},
         )
@@ -2112,6 +2181,7 @@ class PromptFeatureTest {
             store = store,
             tabsUseCases = mock(),
             fragmentManager = fragmentManager,
+            fileUploadsDirCleaner = mock(),
             exitFullscreenUsecase = mock(),
             shareDelegate = delegate,
         ) { }
@@ -2143,6 +2213,7 @@ class PromptFeatureTest {
             activity = mock(),
             store = store,
             tabsUseCases = mock(),
+            fileUploadsDirCleaner = mock(),
             fragmentManager = fragmentManager,
             exitFullscreenUsecase = mock(),
             shareDelegate = delegate,
@@ -2175,6 +2246,7 @@ class PromptFeatureTest {
                 activity = mock(),
                 store = store,
                 tabsUseCases = mock(),
+                fileUploadsDirCleaner = mock(),
                 fragmentManager = fragmentManager,
                 exitFullscreenUsecase = mock(),
                 shareDelegate = mock(),
@@ -2216,6 +2288,7 @@ class PromptFeatureTest {
                 activity = mock(),
                 store = store,
                 tabsUseCases = mock(),
+                fileUploadsDirCleaner = mock(),
                 fragmentManager = fragmentManager,
                 exitFullscreenUsecase = mock(),
                 shareDelegate = mock(),
@@ -2251,6 +2324,7 @@ class PromptFeatureTest {
                 store = store,
                 tabsUseCases = mock(),
                 fragmentManager = fragmentManager,
+                fileUploadsDirCleaner = mock(),
                 exitFullscreenUsecase = mock(),
                 shareDelegate = mock(),
             ) { },
@@ -2280,6 +2354,7 @@ class PromptFeatureTest {
             store = store,
             tabsUseCases = mock(),
             fragmentManager = fragmentManager,
+            fileUploadsDirCleaner = mock(),
             shareDelegate = mock(),
             isSaveLoginEnabled = { true },
             loginValidationDelegate = mock(),
@@ -2319,6 +2394,7 @@ class PromptFeatureTest {
                 activity = mock(),
                 store = store,
                 tabsUseCases = mock(),
+                fileUploadsDirCleaner = mock(),
                 fragmentManager = fragmentManager,
                 shareDelegate = mock(),
             ) { },
@@ -2362,6 +2438,7 @@ class PromptFeatureTest {
                 activity = mock(),
                 store = store,
                 tabsUseCases = mock(),
+                fileUploadsDirCleaner = mock(),
                 fragmentManager = fragmentManager,
                 shareDelegate = mock(),
             ) { },
@@ -2395,6 +2472,7 @@ class PromptFeatureTest {
             store = store,
             tabsUseCases = mock(),
             fragmentManager = fragmentManager,
+            fileUploadsDirCleaner = mock(),
             shareDelegate = mock(),
             isSaveLoginEnabled = { true },
             loginValidationDelegate = mock(),
@@ -2424,6 +2502,7 @@ class PromptFeatureTest {
             store = store,
             tabsUseCases = mock(),
             fragmentManager = fragmentManager,
+            fileUploadsDirCleaner = mock(),
             exitFullscreenUsecase = mock(),
         ) { }
         feature.start()
@@ -2453,6 +2532,7 @@ class PromptFeatureTest {
             store = store,
             tabsUseCases = mock(),
             fragmentManager = fragmentManager,
+            fileUploadsDirCleaner = mock(),
             exitFullscreenUsecase = mock(),
         ) { }
         feature.start()
@@ -2483,6 +2563,7 @@ class PromptFeatureTest {
             tabsUseCases = mock(),
             fragmentManager = fragmentManager,
             exitFullscreenUsecase = mock(),
+            fileUploadsDirCleaner = mock(),
             isCreditCardAutofillEnabled = { true },
             creditCardValidationDelegate = mock(),
         ) { }
@@ -2568,6 +2649,7 @@ class PromptFeatureTest {
             store = store,
             tabsUseCases = mock(),
             fragmentManager = fragmentManager,
+            fileUploadsDirCleaner = mock(),
             isCreditCardAutofillEnabled = { true },
             creditCardValidationDelegate = mock(),
         ) { }
@@ -2632,6 +2714,7 @@ class PromptFeatureTest {
             tabsUseCases = mock(),
             fragmentManager = fragmentManager,
             exitFullscreenUsecase = mock(),
+            fileUploadsDirCleaner = mock(),
             isCreditCardAutofillEnabled = { true },
             creditCardValidationDelegate = mock(),
         ) { }
@@ -2653,6 +2736,7 @@ class PromptFeatureTest {
             store = store,
             tabsUseCases = mock(),
             fragmentManager = fragmentManager,
+            fileUploadsDirCleaner = mock(),
             isCreditCardAutofillEnabled = { true },
             creditCardValidationDelegate = mock(),
         ) { }
@@ -2696,6 +2780,7 @@ class PromptFeatureTest {
                 activity = mock(),
                 store = store,
                 tabsUseCases = mock(),
+                fileUploadsDirCleaner = mock(),
                 fragmentManager = fragmentManager,
                 exitFullscreenUsecase = mock(),
                 shareDelegate = mock(),
@@ -2724,6 +2809,7 @@ class PromptFeatureTest {
         val feature = spy(
             PromptFeature(
                 activity = mock(),
+                fileUploadsDirCleaner = mock(),
                 store = store,
                 tabsUseCases = mock(),
                 fragmentManager = fragmentManager,